Update YazDHCP.sh
Changes to avoid restarting dnsmasq too often: one time for each call to Update_Hostnames(), Update_Staticlist() & Update_Optionslist(). This prevents the situation when dnsmasq is restarted 3 times in a row when adding an IP reservation along with the optional "DNS Server" & "Hostname" entry fields at the same time.

Changes to avoid calling the YazDHCP script every time the "/jffs/scripts/service-event" script is executed. Now we check if the event is actually related to YazDHCP.