Update YazDHCP.sh
Changes/additions to fix a problem that was found during the reboot sequence, when the USB-attached disk drive (where the backup directory path for custom user icons is located) is not yet fully mounted & available during the time when YazDHCP is started up via a call from the "services-start" script. Now the fixed code re-initializes the backup directory path when the add-on is executed by the user on demand *after* the reboot has been completed.